Initial work on Linux support

+++ Coral.Native/Include/Coral/StableVector.hpp
@@ -2,8 +2,11 @@
 
 #include <vector>
 #include <array>
+#include <mutex>
 #include <shared_mutex>
 #include <list>
+#include <atomic>
+#include <cstring>
 
 namespace Coral {
 
@@ -78,7 +81,7 @@ namespace Coral {
 					if (pageIndex >= m_PageCount)
 					{
 						auto oldPages = m_PageCount;
-						m_PageCount = std::max(16ull, m_PageCount * 2);
+						m_PageCount = std::max(uint64_t(16), m_PageCount * 2);
 						auto newPageTable = std::make_unique<Page*[]>(m_PageCount);
 						std::memcpy(newPageTable.get(), m_PageTable.load(), oldPages * sizeof(void*));
 						m_PageTable.exchange(newPageTable.get());
@@ -111,7 +114,7 @@ namespace Coral {
 					if (pageIndex >= m_PageCount)
 					{
 						auto oldPages = m_PageCount;
-						m_PageCount = std::max(16ull, m_PageCount * 2);
+						m_PageCount = std::max(uint64_t(16), m_PageCount * 2);
 						auto newPageTable = std::make_unique<Page*[]>(m_PageCount);
 						std::memcpy(newPageTable.get(), m_PageTable.load(), oldPages * sizeof(void*));
 						m_PageTable.exchange(newPageTable.get());
@@ -140,7 +143,7 @@ namespace Coral {
 				if (pageIndex >= m_PageCount)
 				{
 					auto oldPages = m_PageCount;
-					m_PageCount = std::max(16ull, m_PageCount * 2);
+					m_PageCount = std::max(uint64_t(16), m_PageCount * 2);
 					auto newPageTable = std::make_unique<Page*[]>(m_PageCount);
 					std::memcpy(newPageTable.get(), m_PageTable.load(), oldPages * sizeof(void*));
 					m_PageTable.exchange(newPageTable.get());
@@ -167,7 +170,7 @@ namespace Coral {
 				if (pageIndex >= m_PageCount)
 				{
 					auto oldPages = m_PageCount;
-					m_PageCount = std::max(16ull, m_PageCount * 2);
+					m_PageCount = std::max(uint64_t(16), m_PageCount * 2);
 					auto newPageTable = std::make_unique<Page*[]>(m_PageCount);
 					std::memcpy(newPageTable.get(), m_PageTable.load(), oldPages * sizeof(void*));
 					m_PageTable.exchange(newPageTable.get());
